Refactored and adjusted the behavior of the code parts used to calculate the number of days between two dates. Analytic fields are now dynamically added to the journal lines. Added comments and docstrings.
Both the _calculate_days method and the _generate_depreciations generator used different ways to calculate the number of days in a time interval. This could result in incoherences, as observed in some unusual cases. Those conflicting algorithms were refactored into one method, _nb_days_in_interval, with the added benefit of removing the hardcoding of 30-day months/periods from inside _generate_depreciations.

As a result, the module is also more flexible with the types of periods it can work with. Periods with a length of one month are now supported even when they do not start and end with the month itself. Periods with a (fixed) length of several months are correctly depreciated over by assets that are using the Ending Date method, and it will be much easier to add full support when needed.