Improved the module's behavior with multiple companies.
Implemented domains and constraints to ensure that an asset, its category, and the accounts and journals that correspond to the category are all associated with the same company.

When depreciating over a given period, only assets from the same company as the period should be depreciated. As a consequence, the domain in the depreciation wizard has been modified, and a condition added to the depreciate function. Also updated the behavior of the method that computes the projection, in order to only use periods from the right company.